% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clustering.R
\name{AllZero}
\alias{AllZero}
\title{AllZero}
\usage{
AllZero(g)
}
\arguments{
\item{g}{The input graph}
}
\value{
boolean value indicating whether all active nodes are 0 or not
}
\description{
AllZero
}
