% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clustering.R
\name{DeleteNodes}
\alias{DeleteNodes}
\title{DeleteNodes}
\usage{
DeleteNodes(g, node.id)
}
\arguments{
\item{g}{The input graph}

\item{node.id}{the ID of the node(s)}
}
\value{
graph with node.id deleted
}
\description{
DeleteNodes
}
