% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clustering.R
\name{FindArgMinLinkage}
\alias{FindArgMinLinkage}
\title{FindArgMinLinkage}
\usage{
FindArgMinLinkage(g)
}
\arguments{
\item{g}{The input graph}
}
\value{
ID of the node with min linkage value
}
\description{
Find the node with the minimum linkage value
}
