% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clustering.R
\name{Id2Idx}
\alias{Id2Idx}
\title{Id2Idx}
\usage{
Id2Idx(g, node.id)
}
\arguments{
\item{g}{The input graph}

\item{node.id}{The sequence of node IDs}
}
\value{
node.idx corresponding to node.id
}
\description{
Id2Idx
}
