% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clustering.R
\name{Idx2Id}
\alias{Idx2Id}
\title{IdxId}
\usage{
Idx2Id(g, node.idx)
}
\arguments{
\item{g}{The input graph}

\item{node.idx}{The sequence of node indexes (1-based)}
}
\value{
node.id corresponding to node.idx
}
\description{
IdxId
}
