% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clustering.R
\name{LinkageFunctionInit}
\alias{LinkageFunctionInit}
\title{LinkageFunctionInit}
\usage{
LinkageFunctionInit(g)
}
\arguments{
\item{g}{The input graph}
}
\value{
The same graph with "linkage.value" field
}
\description{
Computes the values of the linkage function on each node of the graph
}
