% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clustering.R
\name{LinkageFunctionUpdate}
\alias{LinkageFunctionUpdate}
\title{LinkageFunctionUpdate}
\usage{
LinkageFunctionUpdate(g, node.id)
}
\arguments{
\item{g}{The input graph}

\item{node.id}{node IDs to be removed from the active nodes}
}
\value{
The same graph with "linkage.value" field updated
}
\description{
Remove the node indexes (1-based) and update the linkage function
}
