% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clustering.R
\name{LinkageVal}
\alias{LinkageVal}
\title{LinkageVal}
\usage{
LinkageVal(g, node.id)
}
\arguments{
\item{g}{The input graph}

\item{node.id}{the ID of the node(s)}
}
\value{
linkage value for node.id
}
\description{
LinkageVal
}
