% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{LoadGraph}
\alias{LoadGraph}
\title{LoadGraph}
\usage{
LoadGraph(M)
}
\arguments{
\item{M}{the upper-triangular block matrix created with ParseMultipartiteGraphList}
}
\value{
a graph structure
}
\description{
Load a matrix or a sparse matrix (Matrix object) and store it into an igraph object
Load also the metadata associated to the nodes (i.e. the vertex sets of the multi-partite graph)
}
