% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clustering.R
\name{MultipartiteBestCluster}
\alias{MultipartiteBestCluster}
\title{MultipartiteBestCluster}
\usage{
MultipartiteBestCluster(g)
}
\arguments{
\item{g}{The input graph}
}
\value{
List containing the graph and the cluster
}
\description{
Algorithm to find the best cluster from the current node selection
}
