% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clustering.R
\name{MultipartiteClustering}
\alias{MultipartiteClustering}
\title{MultipartiteClustering}
\usage{
MultipartiteClustering(g)
}
\arguments{
\item{g}{The input graph}
}
\value{
The list of clusters and their values
}
\description{
Iterative procedure to find all clusters
}
