% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ParseMultipartiteGraphList}
\alias{ParseMultipartiteGraphList}
\title{ParseMultipartiteGraphList}
\usage{
ParseMultipartiteGraphList(file.list)
}
\arguments{
\item{file.list}{txt file containing the matrices}
}
\value{
the upper-triangular block matrix and the block ID for each row/column
}
\description{
Parse a named, ordered list of bipartite graph file names and store it into a 0-diagonal, symmetric block matrix
The matrices must be parsed in row order and then column order
N.B. 1) the # of columns of vertex set i must equate the # of rows of vertex set i+1
     2) rownames, colnames: auto-detected, but not used
}
