% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{PrintMultipartiteClustering}
\alias{PrintMultipartiteClustering}
\title{PrintMultipartiteClustering}
\usage{
PrintMultipartiteClustering(cluster.list, file.name)
}
\arguments{
\item{cluster.list}{The output of MultipartiteClustering}

\item{file.name}{output txt file}
}
\value{
nothing
}
\description{
Print the clusters, one per line
}
