% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_clustering.R
\name{SetNodesInH}
\alias{SetNodesInH}
\title{GetNodesInH}
\usage{
SetNodesInH(g, node.id)
}
\arguments{
\item{g}{The input graph}

\item{node.id}{The sequence of nodes}
}
\value{
the same graph with node.id nodes active
}
\description{
GetNodesInH
}
